requirements.txt
docopt
gitpython
jinja2
six
